[TM-3425] start up polygon tab redisgn for admins
